New hooks

+++ functions/SelectMenu/S_player_Search.js
@@ -1,4 +1,4 @@
-const { useFunctions } = require("@zibot/zihooks");
+const { useHooks } = require("@zibot/zihooks");
 
 module.exports.data = {
 	name: "S_player_Search",
@@ -12,9 +12,17 @@ module.exports.data = {
  */
 
 module.exports.execute = async ({ interaction, lang }) => {
+	// Check if useHooks is available
+	if (!useHooks) {
+		console.error("useHooks is not available");
+		return (
+			interaction?.reply?.({ content: "System is under maintenance, please try again later.", ephemeral: true }) ||
+			console.error("No interaction available")
+		);
+	}
 	const query = interaction.values?.at(0);
 	if (query === "B_cancel") return interaction.message.delete().catch((e) => {});
-	const command = useFunctions().get("Search");
+	const command = useHooks.get("functions").get("Search");
 	await command.execute(interaction, query, lang);
 	return;
 };
